requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SQLAlchemy==3.1.1
GitPython==3.1.44
gTTS==2.5.4
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.8
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
smmap==5.0.2
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
zipp==3.23.0
psycopg2-binary==2.9.10
